refactor: rename pkg/owasp to pkg/orchestrator (LAB-961)
The package contains generic test orchestration infrastructure (mutation
executors, resource trackers, endpoint selectors, template loaders),
none of which is OWASP-specific. Renaming accurately reflects its role
as a coordinator between pkg/runner and pkg/templates.
