+++ pkg/orchestrator/apikey_auth_test.go
@@ -1,4 +1,4 @@
-package owasp
+package orchestrator
 
 import (
 	"context"
