+++ pkg/orchestrator/loader_sort_test.go
@@ -1,4 +1,4 @@
-package owasp
+package orchestrator
 
 import (
 	"os"
