+++ pkg/orchestrator/selector.go
@@ -1,4 +1,4 @@
-package owasp
+package orchestrator
 
 import (
 	"regexp"
